make separate actions for get/post reset password
