+++ src/login/app_config/login-rootSaga.js
@@ -2,14 +2,15 @@ import { takeLatest, takeEvery } from "redux-saga/effects";
 import * as init_actions from "../app_init/init_actions";
 import { requestConfig } from "../app_init/init_sagas";
 import { getResetPasswordConfig, postEmailLink } from "../redux/sagas/resetpassword/resetPasswordSaga";
-import * as resetPasswordActions from "../redux/actions/resetPasswordActions";
+import * as getResetPasswordActions from "../redux/actions/getResetPasswordActions";
+import * as postResetPasswordActions from "../redux/actions/postResetPasswordActions";
 
 function* rootSaga() {
   yield [
     takeLatest(init_actions.GET_CONFIG, requestConfig),
-    takeLatest(resetPasswordActions.GET_RESET_PASSWORD, getResetPasswordConfig),
-    takeLatest(resetPasswordActions.POST_RESET_PASSWORD, postEmailLink),
-    takeEvery(resetPasswordActions.POST_RESET_PASSWORD_FAIL, getResetPasswordConfig),
+    takeLatest(getResetPasswordActions.GET_RESET_PASSWORD, getResetPasswordConfig),
+    takeLatest(postResetPasswordActions.POST_RESET_PASSWORD, postEmailLink),
+    takeEvery(postResetPasswordActions.POST_RESET_PASSWORD_FAIL, getResetPasswordConfig),
   ];
 }
