+++ src/login/redux/reducers/resetPasswordReducer.js
@@ -1,4 +1,5 @@
-import * as actions from "../actions/resetPasswordActions";
+import * as getActions from "../actions/getResetPasswordActions";
+import * as postActions from "../actions/postResetPasswordActions";
 
 const data = {
     email: "",
@@ -7,12 +8,12 @@ const data = {
 
 let resetPasswordReducer = (state = data, action) => {
   switch (action.type) {
-    case actions.GET_RESET_PASSWORD_SUCCESS:
+    case getActions.GET_RESET_PASSWORD_SUCCESS:
       return {
         ...state,
        ...action.payload
       };
-      case actions.POST_RESET_PASSWORD:
+      case postActions.POST_RESET_PASSWORD:
         return {
           ...state,
          ...action.payload
